requirements.txt
langchain==0.1.0
langchain_core==0.1.13
langchain_openai==0.0.2.post1
streamlit==1.31.0
python-dotenv==1.0.0
